src/test/java/fr/elyssif/client/StringUtilsTest.java
/*
 * Elyssif-Client
 * Copyright (C) 2019 Jérémy LAMBERT (System-Glitch)
 *
 * This program is free software: you can redistribute it and/or modify
 * it under the terms of the GNU General Public License as published by
 * the Free Software Foundation, either version 3 of the License, or
 * (at your option) any later version.

 * This program is distributed in the hope that it will be useful,
 * but WITHOUT ANY WARRANTY; without even the implied warranty of
 * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 * GNU General Public License for more details.

 * You should have received a copy of the GNU General Public License
 * along with this program.  If not, see <https://www.gnu.org/licenses/>.
 */
 package fr.elyssif.client;

import static org.junit.jupiter.api.Assertions.assertEquals;

import org.junit.jupiter.api.Test;

class StringUtilsTest {

	@Test
	void testToSnakeCase() {
		String camel = "camelCase";
		String snake = StringUtils.toSnakeCase(camel);
		assertEquals("camel_case", snake);
	}

	@Test
	void testToCamelCase() {
		String snake = "snake_case";
		String camel = StringUtils.toCamelCase(snake);
		assertEquals("snakeCase", camel);
	}

}
